+++ cli/installer/posix.go
@@ -1,0 +1,1 @@
+package installer
